+++ Solutions/Armis/ReleaseNotes.md
@@ -1,5 +1,6 @@
 | **Version** | **Date Modified (DD-MM-YYYY)** | **Change History**                          |
 |-------------|--------------------------------|---------------------------------------------|
+| 3.1.0       | 11-09-2024                     | Updated Armis Alerts Data connector to ingest Armis Activities associated with only Armis Alerts.|
 | 3.0.3       | 26-08-2024                     | Updated the python runtime version to **3.11**|
 | 3.0.2       | 03-05-2024                     | Repackaged for parser issue fix on reinstall|
 | 3.0.1       | 15-04-2024                     | Added Deploy to Azure Government button in **Data connectors**|