+++ Solutions/MarkLogicAudit/ReleaseNotes.md
@@ -1,3 +1,4 @@
 | **Version** | **Date Modified (DD-MM-YYYY)** | **Change History**                                          |
 |-------------|--------------------------------|-------------------------------------------------------------|
+| 3.0.1       | 02-01-2025                     | Removed Deprecated **Data connector**                       |
 | 3.0.0       | 12-08-2024                     | Deprecating data connector                                  |