+++ Solutions/Ubiquiti UniFi/ReleaseNotes.md
@@ -1,5 +1,6 @@
 | **Version** | **Date Modified (DD-MM-YYYY)** | **Change History**                                                        |
 |-------------|--------------------------------|---------------------------------------------------------------------------|
+| 3.0.3       | 04-12-2024                     | Removed Deprecated **Data Connector**                                     |
 | 3.0.2       | 09-08-2024                     | Deprecating data connectors                                               |
 | 3.0.1       | 16-07-2024                     | Updated the **Analytic rules** for missing TTP					   		   |
 | 3.0.0       | 23-01-2024                     | Updated the **Data Connector** by removing preview-tag   				   |